refactor(tasks): extract shared utils and unify message analysis
- Create tasks/utils.ts with canonical toErrorMessage + isPromiseLike
- Remove toErrorMessage duplicates from task-launcher, question-relay, permission-proxy
- Deduplicate ErrorCategory to types.ts only (error-classifier re-exports)
- Add 6 message analysis functions to session-messages.ts (getMessageTime, getLastAssistantMessage, getFinishReason, extractToolCallSequence, hasAssistantTextContent, extractAssistantText)
- Remove duplicate functions from progress-analyzer, loop-detector, idle-diagnostic
- All consumers updated to import from canonical locations
- Tests: 409 passed / 7 skipped (baseline unchanged)
