+++ wopal-plugin/src/tasks/progress-analyzer.ts
@@ -1,5 +1,11 @@
 import type { SessionMessage } from "../types.js"
 import { createDebugLog } from "../debug.js"
+import {
+  getMessageTime,
+  extractToolCallSequence,
+  hasAssistantTextContent,
+  getFinishReason,
+} from "./session-messages.js"
 
 const debugLog = createDebugLog("[wopal-task]", "task")
 
@@ -13,73 +19,6 @@ export interface ProgressInfo {
 }
 
 /**
- * Get timestamp from a message's time field.
- * Returns milliseconds since epoch, or 0 if unavailable.
- */
-function getMessageTime(message: SessionMessage): number {
-  const time = message.info?.time
-  if (!time) return 0
-
-  if (typeof time === "string") {
-    const parsed = Date.parse(time)
-    return isNaN(parsed) ? 0 : parsed
-  }
-
-  return time.created ?? 0
-}
-
-/**
- * Count tool calls from message parts.
- * OpenCode uses part.type === "tool" with part.tool containing the tool name.
- */
-function countToolCalls(messages: SessionMessage[]): Map<string, number> {
-  const toolCounts = new Map<string, number>()
-
-  for (const message of messages) {
-    if (message.info?.role !== "assistant") continue
-
-    for (const part of message.parts ?? []) {
-      if (part.type === "tool") {
-        const toolName = part.tool ?? "unknown"
-        const current = toolCounts.get(toolName) ?? 0
-        toolCounts.set(toolName, current + 1)
-      }
-    }
-  }
-
-  return toolCounts
-}
-
-/**
- * Check if messages contain any assistant text content.
- */
-function hasAssistantTextContent(messages: SessionMessage[]): boolean {
-  for (const message of messages) {
-    if (message.info?.role !== "assistant") continue
-
-    for (const part of message.parts ?? []) {
-      if ((part.type === "text" || part.type === "reasoning") && part.text?.trim()) {
-        return true
-      }
-    }
-  }
-  return false
-}
-
-/**
- * Get the finish reason from the last assistant message.
- */
-function getFinishReason(messages: SessionMessage[]): string | undefined {
-  for (let i = messages.length - 1; i >= 0; i--) {
-    const message = messages[i]
-    if (message.info?.role === "assistant") {
-      return message.info?.finish
-    }
-  }
-  return undefined
-}
-
-/**
  * Analyze progress information from session messages.
  * 
  * @param allMessages - All messages in the session
@@ -90,7 +29,11 @@ export function analyzeProgress(
   allMessages: SessionMessage[],
   newMessages: SessionMessage[]
 ): ProgressInfo {
-  const toolCounts = countToolCalls(allMessages)
+  const toolSequence = extractToolCallSequence(allMessages)
+  const toolCounts = new Map<string, number>()
+  for (const tool of toolSequence) {
+    toolCounts.set(tool, (toolCounts.get(tool) ?? 0) + 1)
+  }
   const toolCalls = Array.from(toolCounts.entries())
     .map(([tool, count]) => ({ tool, count }))
     .sort((a, b) => b.count - a.count)
