+++ wopal-plugin/src/tasks/question-relay.ts
@@ -1,5 +1,6 @@
 import type { SimpleTaskManager } from "./simple-task-manager.js"
 import { createDebugLog, type DebugLog } from "../debug.js"
+import { toErrorMessage } from "./utils.js"
 
 const defaultDebugLog = createDebugLog("[wopal-task]", "task")
 
@@ -118,22 +119,4 @@ This question requires your attention. The background task is waiting.
     log(`[question] notify parent failed for task ${taskId}: ${toErrorMessage(err)}`)
     throw err // Re-throw to be caught by caller
   }
-}
-
-function toErrorMessage(error: unknown): string {
-  if (error instanceof Error && error.message) {
-    return error.message
-  }
-  if (typeof error === "string" && error.length > 0) {
-    return error
-  }
-  try {
-    const serialized = JSON.stringify(error)
-    if (serialized && serialized !== "{}") {
-      return serialized
-    }
-  } catch {
-    // Ignore JSON serialization failures
-  }
-  return String(error)
 }
